requirements.txt
requests==2.31.0
feedparser==6.0.10
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
